Added changes in Payment, Inventory and Vend Machine

+++ src/main/java/com/techelevator/Chips.java
@@ -1,8 +1,10 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Chips extends Product{
 		
-	public Chips(String name, double price) {
+	public Chips(String name, BigDecimal price) {
 		super(name, price);
 	}
